Rename/publishing: HomeBody1, Body2 순서 맞추기
